refactoring (ai/rsc): extract render into separate file
