+++ packages/core/rsc/rsc-server.ts
@@ -1,8 +1,5 @@
 export { getAIState, getMutableAIState } from './ai-state';
-export {
-  createStreamableUI,
-  createStreamableValue,
-  render,
-} from './streamable';
+export { createStreamableUI, createStreamableValue } from './streamable';
+export { render } from './render';
 export { streamUI } from './stream-ui';
 export { createAI } from './provider';
