+++ packages/core/rsc/streamable.ui.test.tsx
@@ -3,12 +3,7 @@ import {
   openaiFunctionCallChunks,
 } from '../tests/snapshots/openai-chat';
 import { DEFAULT_TEST_URL, createMockServer } from '../tests/utils/mock-server';
-import {
-  createStreamableUI,
-  createStreamableValue,
-  render,
-} from './streamable';
-import { z } from 'zod';
+import { createStreamableUI, createStreamableValue } from './streamable';
 
 const FUNCTION_CALL_TEST_URL = DEFAULT_TEST_URL + 'mock-func-call';
 
@@ -187,71 +182,6 @@ function createMockUpProvider() {
     },
   } as any;
 }
-
-describe('rsc - render()', () => {
-  it('should emit React Nodes with sync render function', async () => {
-    const ui = render({
-      model: 'gpt-3.5-turbo',
-      messages: [],
-      provider: createMockUpProvider(),
-      functions: {
-        get_current_weather: {
-          description: 'Get the current weather',
-          parameters: z.object({}),
-          render: () => {
-            return <div>Weather</div>;
-          },
-        },
-      },
-    });
-
-    const rendered = await simulateFlightServerRender(ui as any);
-    expect(rendered).toMatchSnapshot();
-  });
-
-  it('should emit React Nodes with async render function', async () => {
-    const ui = render({
-      model: 'gpt-3.5-turbo',
-      messages: [],
-      provider: createMockUpProvider(),
-      functions: {
-        get_current_weather: {
-          description: 'Get the current weather',
-          parameters: z.object({}),
-          render: async () => {
-            await new Promise(resolve => setTimeout(resolve, 100));
-            return <div>Weather</div>;
-          },
-        },
-      },
-    });
-
-    const rendered = await simulateFlightServerRender(ui as any);
-    expect(rendered).toMatchSnapshot();
-  });
-
-  it('should emit React Nodes with generator render function', async () => {
-    const ui = render({
-      model: 'gpt-3.5-turbo',
-      messages: [],
-      provider: createMockUpProvider(),
-      functions: {
-        get_current_weather: {
-          description: 'Get the current weather',
-          parameters: z.object({}),
-          render: async function* () {
-            yield <div>Loading...</div>;
-            await new Promise(resolve => setTimeout(resolve, 100));
-            return <div>Weather</div>;
-          },
-        },
-      },
-    });
-
-    const rendered = await simulateFlightServerRender(ui as any);
-    expect(rendered).toMatchSnapshot();
-  });
-});
 
 describe('rsc - createStreamableUI()', () => {
   it('should emit React Nodes that can be updated', async () => {
